Moved methods into their appropriate classes
